// submodules/common-dto/api-client/main/models/StyleDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StyleDto = {
    /**
     * Id of style
     */
    id: number;
    /**
     * Slug
     */
    slug: string;
    /**
     * Localized title of style
     */
    title: string;
    /**
     * ID of image
     */
    image_id: string | null;
};